replace packages command with scan
